requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pyyaml>=6.0.1

# ── HTTP / async ──────────────────────────────────────────────────────────
httpx>=0.27.0

# ── Retry / resilience ────────────────────────────────────────────────────
tenacity>=8.3.0

# ── SerpAPI ───────────────────────────────────────────────────────────────
google-search-results>=2.4.2

# ── Fuzzy deduplication ───────────────────────────────────────────────────
thefuzz>=0.22.1
python-Levenshtein>=0.25.0     # accelerates thefuzz

# ── Embedding model (local inference) ────────────────────────────────────
sentence-transformers>=3.0.0
torch>=2.2.0                   # required by sentence-transformers

# ── CLI progress ──────────────────────────────────────────────────────────
tqdm>=4.66.0

# ── Testing ───────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
